Updating from comments
